Improve pref tests logging info, group duplicate codes to common.py
